linux: Example and Demos: remove Cryptography section
Based on LCPD-12405, the crypto demos will be removed from the root
filesystem, so the first portion of the Cryptography section which
describes the demos should be removed.

The second portion is out of date, and redundant to
Foundational_Components/Kernel/Kernel_Drivers/Crypto.rst.inc

so remove the whole Cryptography section from Example and Demos.

Signed-off-by: Bin Liu <b-liu@ti.com>