HSR/PRP: Update for Dual instance/SV frame with VLAN tag
Modify the User guide and Migration guide for new features added to
Processor SDK 5.1 such as Dual instance support for HSR/PRP and VLAN
tag support in SV frames. Modify Migration guide to update the changes
to debugfs file paths and file names as applicable and provide details
of procfs replacements for the same.

Additionally feature list is added.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>