linux: update ADC kernel driver docs
The existing code snippet to enable the ADC driver does not actually enable
the driver, since tscadc is disabled in the initial dtsi file. This patch adds
status = "okay"; to the sample device tree file code, and a link to TI's
documentation on building and installing kernel modules.

This patch is a closed-loop action based on CATAPPS-691.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>