common: link to PRU header information
There is much more detailed information on PRU header files in the
PRU Optimizing C/C++ Compiler User's Guide than there is in the SDK docs.
This commit adds notes that point customers to the right parts of the User's
Guide if they want to learn more about header files.

It also updates a broken link in the Getting Started with PRU-ICSS section that
should point to the Header Files section.

Ties to closed loop action CATAPPS-507

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>