U-Boot: Remove AM335x SPI boot related documentations
There is no requirement to support SPI Boot mode in U-Boot on AM335x and
therefore not tested. This feature is broken for quite sometime and
hence remove it from user guide so as not to mislead users. Also
indicate that SPI boot mode is not supported in toplevel table

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>