linux: Add patches for using Cypress S28 flash on J721E and vice versa
Flashes can be interchanges quite easily with small modifications to the
read-delay and PHY tuning pattern location. Add patches that show how to
use the Cypress S28 flash on J721E and Micron MT35 flash on J7200.

The patches are only compile-tested because I don't have access to these
combination of boards/flashes. But I have used the Cypress flash on
J721E earlier with the corresponding patch.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>