AM64x_config Update
Updated the version of Linux the SDK was tested on (Ubuntu 18.4 64 bit only)

Updated the version of Windows the SDK was tested on to 10 (64 bit).

Updated the installer filename. Note that this is what we expect the installer
filename to become in the future. The initial installer filename does not match
the pattern that we set with previous devices.

TODO: are we expecting the [variables] to be replaced? (e.g., for [platform] to
get replaced by am64xx-evm in AM64x builds?) Because no replacements
are occurring.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>