U-Boot: qspi: Add a binary file containing the PHY pattern
For PHY calibration to succeed, the pattern must be flashed first. While
the instructions to flash are present, the actual pattern is not. Add
the pattern file so users can download it and flash it. Update the file
name in the instructions to match the name of the file containing the
pattern.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>