[PDK-10335] Documentation fix and cleanup for AM57x.
Steps are already captured to run the example, so remove the TODO part.
Fix the file where links are not redirecting properly.
Change the SYS_BIOS link which is visible by outside TI network.
Removed AM65 IPC example from the content bar as it is a part of CoreSDK RTOS(PDK).

Signed-off-by: Jayanth BR <x1080850@ti.com>