UG-Memory: Document the resolution for failure of USB storage device detection failure in U-Boot
In U-Boot and while booting from a USB Mass storage device, some devices
fail to get enumurated, this is because these fail to follow the spec for
power good delay. This issue can be resloved setting the environment
varaible 2000, which is seems to be good enough for all cases. This can not
be implemented by default as this would slow the enumeration for devices
that follow the spec too.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>