PRU-ICSS/Linux_Drivers/PRU_ICSSG_Ethernet: fix ifconfig usage
The "ifconig" has limitation "This directive is designed to control only
content of document. It could not control sections, labels and so on.". As
result, section add by below commit will not be displayed as it contains
section definition.

Fix it by moving section header out and add note for AM64X that TSN is not
supported yet.

Fixes: 4163cf8c ("linux: PRU_ICSSG: Add TSN related doc for ICSSG")
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>