configs: J7200/DRA821: linux_toc: Remove DSS config
Remove DSS7 configuration from the J7200 and DRA821A SoCs. These SoCs do
not support the display subsystem.

However, running the check-warn script will show a warning, because
a mismatch is created between the config and the toctree. The J7200
toctree file will have to be re-implemented the way its done for AM64x
SoC.

code-block::
$ ./bin/check-warn.sh -o linux -a <old-head-id> -b <new-head-id> -d J7200

<path-to>/processor-sdk-doc/source/linux/Foundational_Components_Kernel_Drivers.rst:4: WARNING: toctree contains reference to nonexisting document u'linux/Foundational_Components/Kernel/Kernel_Drivers/Display/DSS7'

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>