linux: How_to_Guides: Add instructions to enable OLDI display on AM62 SK
The OLDI panel DT overlay file for AM62 SK has not been added to the wic
image for the 8.4 release. These instructions shall be reduced to
overlay enablement in u-boot only in the next release as the dtbo file
will be added to the wic image.

Add instructions to build it from the PSDK linux source tree and add it
in the uEnv.txt.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>