Kernel_Drivers/Network/CPSWng: Add documentation for CPSWng Native Ethernet
Add new document for CPSWng subsystem in order to describe the two modes
of configuration of the CPSWng subsystem:
1. Virtual MAC
2. Native Ethernet

Add documentation for the Native Ethernet support present in Jacinto
devices.

Move documentation for "CPSWng Virtual MAC" and "CPSWng Native Ethernet"
within the "CPSWng" document. This change is required to ensure a single
source of information for the CPSWng subsystem in order to avoid
confusion.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>