Kernel: MCAN: Add documentation for AM62x SoC
AM62x SK has 3 instances of MCAN on the SoC. CAN driver is enabled
on AM62x, but the board does not carry on-board CAN-FD transceivers.
Update the documentation to reflect this.

The following commit will include a guide that shows how to connect
a CAN external transcever to AM62x SK to test CAN on AM62x.

Signed-off-by: Judith Mendez <jm@ti.com>