Linux: Foundational Components: OP-TEE TF-A instructions and bug fixes
Fix code block from :: to .. code-block:: <language>.

Update build instructions. Since OPTEE build requires both 32 bit and 64
bit compiler, update instructions to reflect this.

Update OP-TEE TF-A commits to now point to release notes to better maintain
which versions of TF-A and OPTEE we are using for each release.

LCPD-34342

Signed-off-by: Judith Mendez <jm@ti.com>