linux: Foundational Components: Fix code block instances
 - Kernel users guide
 - U-boot general info
 - U-boot Splash screen
 - PSDK Directory Structure

Code blocks are not being displayed. Change from ".. code-block::"
to ".. code-block:: <language>" so all code blocks in U-boot and Linux Kernel
general guides will be displayed. Change ":: " to ".. code-block:: <language>"
while were at it as well.

Change "U-Boot #" with "=>" for u-boot console command to match the rest
f the documentation. Add # before the comments as well.

Signed-off-by: Judith Mendez <jm@ti.com>