AM64x: Restructure the docs
For some reason, the makefile and scripts are written in such a way that
it handles AM64x differently than other platforms. OS argument is
ignored and the only available config (Linux) is always used.

We have to add Debian documentation for AM64x which is not possible with
the current implementation of build scripts.

This commit updates the directory structure, config files and scripts to
add support for the same.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>