ATF: Fix build commands
Build for TF-A with Low Power support was introduced but the title
is not within the if statement, therefore it will always be printed.
Fix so that TF-A build with LPM support only shows for AM62x and AM62a
platforms.

For AM62p, two build commands for TF-A are displayed so fix this as well.

fixes: d1609e4e46 (ATF: AM62XX : Add build command for ATF Low power modes)
fixes:3f05fd15aa (ATF: Update build command to include flag needed for low power modes)
Signed-off-by: Judith Mendez <jm@ti.com>