_templates: add the ti header back in
For some reason the submodule split the ti header into it's own file and
did not patch the RTD theme to include it by default. Pull it in
manually using the layout override.

This will need to be revisited as the current CSS padding introduced for
the header exists on mobile even though the header behaves differently
in that layout. Still usable, just ugly. Not a web dev though so I'll
need some assistance here.

Signed-off-by: Randolph Sapp <rs@ti.com>