replacevars.py: fix empty replacement definitions
Currently empty strings are represented with a backslash in the family
config files. I don't know who made that decision but I don't think it's
worth changing it right now. Unfortunately this character if given to
the replacement directive will result in parsing errors.

Because almost everything inherits _replacevars.rst this means that a
lot of other errors end up being masked by the initial parsing error.

Fix it so that empty strings are replaced with a unicode null character
instead of an escaped newline. Rendering wise this results in what I
guess to be the expected behavior (nothing, not even a space is
displayed) but it doesn't generate parsing errors.

Signed-off-by: Randolph Sapp <rs@ti.com>