linux: Foundational_Components: U-Boot: UG-DDRSS: Add multi-bit and multiple single-bit ECC 'ddrss' test
ECC detects multi-bit and multiple single-bit errors which is treated
as an uncorrectable error and there are three different ways to
handle them. The 'ddrss' command can be used to simulate the ECC error
by generating bit errors in SDRAM data.

So, add instructions to simulate ECC multi-bit and multiple single-bit
errors and add ways to handle them.

Signed-off-by: Santhosh Kumar K <s-k6@ti.com>