Kernel/Kernel_Drivers/Network/HSR: Update HSR offload documenation
The steps in the document works fine when executed for the first
time. When executed for the second time and above, some of the
states such as multicast addresses are retained. So update the
steps to address this.

Also adopt the code-block directive, and update the script
sectin after passing through shellcheck

Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>