AM62P: ti-apps-launcher: Add Camera section and fix numbering and images
ti-apps-launcher in 09.02 has a new Camera app for AM62P platform.
This new app replaces LiveCamera and Video Playback apps. So update the
docs accordingly.

Also, specifying both height and width for images is streching the
images. So specify only height to keep the images consistently scaled
throughout the doc.

And, Sections that start inside a ifconfig are numbered wrong. So for
sections like ARM Analytics and 3D Demo, Add the heading for all,
content for the supported platforms and a note for platforms which
doesn't support it.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>