Power_Management: fix toctree entries
Specifying a directory in a toctree does not work. You must point to a
landing page that links to the appropriate subpages. This landing page
was already created but it was in the wrong directory with a
non-standard name. Fix the toc trees and move the landing page to the
proper location.

Signed-off-by: Randolph Sapp <rs@ti.com>