Demo_User_Guide: Update Seva Store Docs for 9.2 SDK Release
- With GPU accelerated Chromium Browser integrated in SDK 9.2 release [1],
seva-launcher-aarch64 will be responsible to create a localhost page on
port 8000 so that Chromium Browser just launches the localhost page once user
clicks Seva Store on ti-apps-launcher. no-browser is set to true so as
to avoid a browser launch [2]. Hence, add a section on How to Launch Seva Store
on host PC Browser.

- The Firefox Browser Container is now not packaged inside the filesystem [3]. Hence, launching
Seva Store via CLI shouldn't be encouraged to customers as pulling the Firefox Container &
then launching Seva Store will not be a good user experience.

- Avoid using width for images & fix height to 400.

- This commit also fixes couple of documentation issues/typos from past.

[1]: https://git.ti.com/cgit/arago-project/meta-arago/commit/?h=kirkstone&id=1c7e65c70d3d0648274b9945a94dc8a3b69e5b94

[2]: https://github.com/TexasInstruments/seva/blob/main/seva-launcher/seva-launcher.go#L30

[3]: https://git.ti.com/cgit/ti-sdk-linux/meta-tisdk/commit/?h=kirkstone&id=70dce5ff45bf02fd8fdcf8ef9d7fd55f5e1aa6b0

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>