all: move to sphinx_rtd_theme with overrides
Our internal theme in swtoolsdev has been trailing behind upstream.
Previously this has resulted in rendering errors with bulleted lists and
odd formatting on mobile layouts. Now it's resulting in rendering errors
in tables. I'm done with it.

We were only overriding a small part of the navigation menu and footer
anyway. This can be done with a template and css override file without
creating a whole fork of the sphinx_rtd_theme theme.

Signed-off-by: Randolph Sapp <rs@ti.com>