How-To Guide: Add Boot Optimization Guide
This is the definitive guide for customers to optimize their boot time
in the Sitara MPU AM62xx portfolio. It links all the work done from
ROM, bootloaders (SBL), kernel and userspace to get a display up in less
than 1s.
SPL optimization are have been moved to a later release and will be
added to the guide then.

Several files need to be added to enable customers achieve these
optimizations faster and reduce the burden on E2E support

Signed-off-by: Aashvij Shenai <a-shenai@ti.com>