CPSW: fix the include section hierarchy
Found a slick trick for making sure included files don't randomly change
the structure of the page they are included from -- send them to the rst
parser on the own and inject the formatted contents into the page.

Currently if a document with section hierarchy 0 is injected in the
middle of a page using hierarchy 1 any sections below the include
directive will treated as a subsection of the section at hierarchy 0.
This mangles the TOC tree and results in unnecessary depth.

The following shows how section depth can be simplified by using the
parser flag of the include directive to force isolated preprocessing.
This means that most (if not all) include files can now sanely start
their sections at hierarchy 0 as well.

https://docutils.sourceforge.io/docs/ref/rst/directives.html#including-an-external-document-fragment

Signed-off-by: Randolph Sapp <rs@ti.com>