Program_MMC_boot_media: fix witespace and lists
Fix mixed indentation depths, trailing whitespace and such other
oddities.

A note to all users: multi-depth bulleted lists cannot be used with
ifconfig. The directive is not expanded before the list is parsed. This
means any list in an ifconfig directive is a completely separate from
any surrounding list entities.

This page should be restructured to avoid using lists entirely because
of this, but I've gone ahead and used sections to split the steps and
kept the list. By avoiding going beyond depth 0 it should at least
*look* like a cohesive block even if the html suggests otherwise...

Signed-off-by: Randolph Sapp <rs@ti.com>