AM62X/AM62PX: Android: Document experimental U-Boot 2024.04
For SDK 10.0, we are still shipping U-Boot 2023.04 to make sure that
we are stable.

In SDK 10.1 we will ship U-Boot 2024.04, which has completely transformed the
bootflow: we move from a custom script to upstream stdboot (bootmeth_android).

Give the user the possibility to try the new bootloaders using a build flag.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>