AM62X/AM62PX: Android: Rework bootloaders build section
The "Bootloader components" build section is quite messy. It has grown over time
and it's difficult to follow.

U-Boot's upstream documentation [1] is way better. Re-model our section to be
closer to upstream doc and add various links to the upstream doc.

This prepares the arrival for building 2 set of bootloaders.

[1] https://docs.u-boot.org/en/latest/board/ti/k3.html#build-procedure
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>