AM62X/AM62PX: Android: Reduce .. include:: usage
We have quite some files in the Android documentation that are
just contain an .. include:: directive.

In many places, the included file is only included once.
This is confusing, as we have 2 files for one content.

Move the content of these included files into the files themselves
to reduce the amount of .. include:: directives.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>