AM62X/AM62PX: Android: Move kernel to dedicated page
The overview section is quite overwhelming: we document how to
fetch and build Android, the Kernel and the Bootloaders.

New users building the SDK for the EVM do not need to
rebuild the kernel nor the bootloaders.

Create a new section, named "Foundational Components" and move
details for the Kernel there.

The bootloaders will be moved in a future patch.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>