Makefile: remove dangerous glob
If BUILDDIR ever evaluates to empty string the current glob pattern will
attempt to recursively remove everything under '/'.

That shouldn't happen as BUILDDIR is set explicitly in the Makefile, but
it is not strongly set, allowing people to shoot themselves in the foot
with external values. This still keeps that footgun, but at least it
won't do anything bad if BUILDDIR is unset.

Signed-off-by: Randolph Sapp <rs@ti.com>