fix: How_to_Guides: Flash eMMC devices
Existing page for flashing eMMC devices is out of date... It references
a script that is no longer packaged in the Linux SDK Installer. Moreover
the script usage may have differed across Jacinto/Sitara SDKs.

Add a new page which serves the same purpose but does not use any such
script and hence is easier to maintain.

Update the TOC to reflect this change for applicable platforms.
AM68 does not have an eMMC. Remove the page from it's TOC.

Signed-off-by: Aniket Limaye <a-limaye@ti.com>