fix: How_to_Guides: M2CC3301: Add separate code-block for Debian
To connect to wifi, a set of scripts must be run. These scripts are
given in the How_To_Enable_M2CC3301_in_linux guide. The syntax is
`./<script-name> <args>`.

However in Debian, it is necessary to prepend `bash` to the command:
`bash ./<script-name> <args>`

This is an upstream scripts' issue. For now, add a separate code-block
for the Debian-specific case. Revert this commit when the issue is
fixed.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>