fix: Kernel: E5010: Add Information on Cropping
Cropping support has been added to the v4l2jpegenc element allowing for
custom crop rectangles to be set. This feature is specific to TI's AM62a
SDK offering as this feature is not in upstream version of Gstreamer.
As a result, information and examples are needed in documentation to
provide proper usage of the property.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>