fix(linux): Power_Management: Fix units of CPU resume latency constraint
One way to enter MCU Only mode is by setting the CPU resume latency
constraint to be 100ms or less. The PM QoS framework is used to set the
CPU resume latency in Linux and the units it uses are us. While the DM,
which recieves the CPU resume Latency constraint, uses ms for the units.
Previous to this SDK, users sent latency in ms to PM QoS. In this
release, it is fixed so that PM QoS receives latency in us.

Signed-off-by: Kendall Willis <k-willis@ti.com>