feat(power-management): Add confirmation of wakeup pin and mode
When the SoC resumes in SDK 11.00 the pin that caused the wakeup event
and the low power mode last entered are printed upon resume of the
system. Explain how to find what wakeup event is associated with the pin
printed and what low power mode was entered. Add an example of finding
this.

Signed-off-by: Kendall Willis <k-willis@ti.com>