fix(power-management): Remove remoteproc device constraint example
Remove the example of how to set a device constraint that leaves the
device powered on. This example references the remoteproc driver. In
this driver, the constraint is being set during suspend_late. At that
point in the suspend sequence, no new constraints are being sent to the
DM so this code does not show a device constraint being set correctly.

Signed-off-by: Kendall Willis <k-willis@ti.com>