feat(graphics): move to index file structure
This renames the Foundational_Components_Graphics landing page to index
and moves it under Foundational_Components/Graphics so it matches the
toctree structure.

Sphinx prefers the concept of indexes, landing pages that will break of
into additional sections and subsections. We've been using that
structure without actually calling them that or placing them in the
proper location.

Signed-off-by: Randolph Sapp <rs@ti.com>