chore: move to exclude only logic based on toc.txt
Exclude all documents with an rst suffix not explicitly listed in the
toc.txt. This removes a lot of dead logic and unnecessary config values.

For a while this project mixed including and excluding document logic,
where the include list was provided as an exhaustive description of
everything through the toc.txt and a basic exclude list was provided as
a sparse array populated manually in the OS tags.py file.

This sparse array was supposed to change the way we searched for index
files (previously referred to as toctreefiles), but ultimately any index
files we needed were already listed in the main toc.txt.

Signed-off-by: Randolph Sapp <rs@ti.com>