chore: add vale config for prose and spell checking
Red Hat has a pretty good template for this thing [1] and it already
understands RST. Might as well pick it up here. Fair warning, it's much
more picky than I am. I'm not entirely sure if we should add it to the
PR checker right now, as even my stuff was flagged to hell and back.

[1] https://github.com/redhat-documentation/vale-at-red-hat

Signed-off-by: Randolph Sapp <rs@ti.com>