docs(linux): U-Boot: Falcon-Mode: Add entry for the DT
The entry node defines the entry point for a binary blob in a fitImage.
Even though the value is not relevant to the device-tree, ommiting it
causes a non-fatal error log in debug u-boot build as follows:

Can't get 'entry' property from FIT

Therefore, this patch adds the entry node which also makes the fitImage
generated using this its file consistent with one's generated during the
yocto build process.

Signed-off-by: Anshul Dalal <anshuld@ti.com>