feat(debian): CC33x1: Update docs to use NetworkManager
Update the CC33x1 "HowTo Guide" to use NetworkManager service and its
nmcli tool for connecting to wifi, instead of using the current method
of cc33xx-target-scripts.

The scripts are required only in Arago, which is a custom distribution.
But on Debian, we should use Debian's default way of managing wifi.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>